+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 Version 1.6-SNAPSHOT
 -------------
 
+CHANGED:
+
+- renamed constant UNDERSCORE_PREFIX to UNDERSCORE in enum DatabasePrefix
+- renamed enum variable UNDERSCORE to UNDERLINE in enum DatabasePrefix
 
 Version 1.5
 -------------